Add logs for the tests of the L2 cache
All logs for the L2 cache are of the form 'LLC_#modelcore#.txt' where
the first # is the number of models run simultaneously and the second #
represents the core(s) used by the model. For example,
LLC_1modelcore0.txt is the log of the single model test where the model
ran on core 0.