Add data sets for PALLOC and new model tests
Added 4 datasets that are split into two directories each (but share the
same platform name in both directories).  The first part of each dataset
was added to the test-model directory and the other part was added to
the test-model-and-benchmark directory, which are both found in the logs
folder.

The datasets added are as follows:
*RPi3OldModel : Tests run with the old model and without PALLOC
*RPi3NewModel : Tests run with the new model and without PALLOC
*RPi3PALLOCOldModel: Tests run with the old model and with PALLOC
*RPi3PALLOCNewModel: Tests run with the new model and with PALLOC

Minor note: replaced the bandwidth executable as the old one was 64-bit.
It should now be a 32-bit executable that can run on the Pi3.