Add scripts that use RT priority
The new scripts are identical to their original counterparts except that
they add RT priority (sudo chrt -f 99) to any command where a test-model
file is called.

Also updated test-model_timings_Pi.sh to use the different models, and
replaced the bandwidth executable, which was 64-bit, with a 32-bit
counterpart so that it can run on the platforms.